Move dxgi.rs to texture
