+++ src/texture/mod.rs
@@ -1,6 +1,7 @@
 mod headers_pc;
 mod headers_ps;
 mod headers_xbox;
+mod dxgi;
 
 pub mod texture;
