+++ src/texture/swizzle/swizzle_xbox.rs
@@ -2,7 +2,7 @@
 
 use log::warn;
 
-use crate::gui::dxgi::XenosSurfaceFormat;
+use crate::texture::dxgi::XenosSurfaceFormat;
 
 use super::Deswizzler;
